requirements.txt
streamlit
Authlib
python-dotenv
pydantic
plotly